third_party/WebKit/LayoutTests/fast/events/touch/gesture/touch-gesture-scroll-page-zoomed-expected.txt

This tests scroll gesture event scrolling on a whole page with browser zoom.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS successfullyParsed is true

TEST COMPLETE
first gesture
PASS document.scrollingElement.scrollTop is 140
PASS document.scrollingElement.scrollLeft is 0
PASS wheelEventsOccurred is 0
second gesture
PASS document.scrollingElement.scrollTop is 40
PASS document.scrollingElement.scrollLeft is 0
PASS wheelEventsOccurred is 0
scroll event 0+> [object HTMLDocument]
PASS scrollEventsOccurred is 1
PASS successfullyParsed is true

TEST COMPLETE
